refacroring limits and reports
